shouroom/requirements.txt
selenium
psycopg2-binary
webdriver-manager
aiogram
beautifulsoup4
lxml